moves SideMenu to general components
